Moved clap definitions to module `cli`
